일단 올리기

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -79,6 +79,8 @@ public class ApiV1PostController {
                 ()-> new ServiceException("401-1", "유효하지 않은 API 키입니다.")
         );
 
+        //인증
+
         Post post = postService.write(actor, reqBody.title, reqBody.content);
         long postsCount = postService.count();
 
@@ -114,10 +116,24 @@ public class ApiV1PostController {
     @Transactional
     public RsData<PostModifyResBody> modify(
             @PathVariable int id,
-            @RequestBody @Valid PostModifyReqBody reqBody
+            @RequestBody @Valid PostModifyReqBody reqBody,
+            @RequestHeader("Authorization") String apiKey
     ) {
 
-        Post post = postService.modify(id, reqBody.title, reqBody.content);
+        apiKey = apiKey.replace("Bearer ", "");
+
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(
+                () -> new ServiceException("401-1", "유효하지 않은 API 키입니다.")
+        );
+
+        Post post = postService.findById(id).get();
+
+        if (!actor.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "수정 권한이 없습니다.");
+        }
+
+
+        postService.modify(id, reqBody.title, reqBody.content);
 
         return new RsData<>(
                 "%d번 게시물이 수정되었습니다.".formatted(post.getId()),
@@ -131,8 +147,20 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Operation(summary="글 삭제")
     public RsData<Void> delete(
-            @PathVariable int id
+            @PathVariable int id,
+            @RequestHeader("Authorization") String apiKey
     ) {
+
+        apiKey = apiKey.replace("Bearer ", "");
+        Member actor = memberService.findByApiKey(apiKey).orElseThrow(
+                () -> new ServiceException("401-1", "유효하지 않은 API 키입니다.")
+        );
+
+        Post post = postService.findById(id).get();
+
+        if(!actor.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "삭제 권한이 없습니다.");
+        }
 
         postService.deleteById(id);
